+++ R/render_xlsx.R
@@ -1,52 +1,46 @@
-require_xlsx <- function() {
-  if (!requireNamespace("xlsx", quietly = TRUE)) {
-    stop("Please install the xlsx package in order to export to excel")
-  }
-}
-
 #' Writes the table to an Excel workbook
 #'
 #' @param x A condformat_tbl object
 #' @param filename The xlsx file name.
 #' @param sheet_name The name of the sheet where the table will be written
-#' @param overwrite_wb logical to overwrite the workbook file
+#' @param overwrite_wb logical to overwrite the whole workbook file
 #' @param overwrite_sheet logical to overwrite the sheet
 #' @export
 #'
 condformat2excel <- function(x, filename, sheet_name = "Sheet1",
                              overwrite_wb = FALSE,
                              overwrite_sheet = TRUE) {
-  require_xlsx()
 
   if (!grepl(pattern = '\\.xlsx$', filename)) { # endsWith(filename, ".xlsx")
     filename <- paste0(filename, ".xlsx")
   }
 
   if (file.exists(filename) && identical(overwrite_wb, FALSE)) {
-    wb <- xlsx::loadWorkbook(filename)
+    if (identical(overwrite_sheet, FALSE)) {
+      stop(paste0(
+        "File ", filename, " already exists. Set overwrite_wb=TRUE ",
+        "to overwrite the whole workbook or overwrite_sheet=TRUE to overwrite ",
+        "just the the sheet ", sheet_name))
+    }
+    wb <- openxlsx::loadWorkbook(filename)
   } else {
-    wb <- xlsx::createWorkbook(type = "xlsx")
+    wb <- openxlsx::createWorkbook(creator = "")
   }
 
-  # getSheets cat's a message I don't want nor care about.
-  noSheets <- wb$getNumberOfSheets()
-  if (noSheets > 0) {
-    sheet_list <- xlsx::getSheets(wb)
-  } else {
-    sheet_list <- list()
-  }
-  if (sheet_name %in% names(sheet_list)) {
+  if (sheet_name %in% names(wb)) {
     if (overwrite_sheet) {
-      xlsx::removeSheet(wb = wb, sheetName = sheet_name)
-      sheet <- xlsx::createSheet(wb, sheetName = sheet_name)
-    } else {
-      sheet <- sheet_list[[sheet_name]]
+      openxlsx::removeWorksheet(wb = wb, sheet = sheet_name)
+      openxlsx::addWorksheet(wb, sheetName = sheet_name)
     }
   } else {
-    sheet <- xlsx::createSheet(wb, sheetName = sheet_name)
+    openxlsx::addWorksheet(wb, sheetName = sheet_name)
   }
-  condformat2excelsheet(x, sheet)
-  xlsx::saveWorkbook(wb, file = filename)
+  condformat2excelsheet(x, wb, sheet_name)
+  openxlsx::saveWorkbook(
+    wb,
+    file = filename,
+    overwrite = overwrite_wb | overwrite_sheet
+  )
   return(invisible(x))
 }
 
@@ -57,45 +51,79 @@ condformat2excel <- function(x, filename, sheet_name = "Sheet1",
 # @examples
 # \dontrun{
 # x <- condformat(iris[1:5,])
-# library(xlsx)
-# wb <- xlsx::createWorkbook(type = "xlsx")
-# sheet <- xlsx::createSheet(wb, sheetName = "Sheet1")
-# condformat2excelsheet(x, sheet)
-# xlsx::saveWorkbook(wb, file = "iris.xlsx")
+# library(openxlsx)
+# wb <- openxlsx::createWorkbook(creator = "")
+# openxlsx::addWorksheet(wb, sheetName = "sheet name")
+# condformat2excelsheet(x, wb, "sheet name")
+# openxlsx::saveWorkbook(wb, file = "iris.xlsx")
 # }
-condformat2excelsheet <- function(x, sheet) {
-  require_xlsx()
-  if (!"jobjRef" %in% class(sheet)) {
-    stop("sheet must be an jobjRef object, as the one returned with xls::createSheet()")
+condformat2excelsheet <- function(x, wb, sheet_name) {
+  xlsx_supported_rules <- c("rule_fill_discrete", "rule_fill_gradient",
+                            "rule_fill_gradient2", "rule_text_bold", "rule_text_color")
+
+  # Check for unsupported rules and warn accordingly:
+  rules <- attr(x, "condformat")[["rules"]]
+  rules_used <- unlist(
+    lapply(rules, function(x) setdiff(class(x), "condformat_rule"))
+  )
+  rules_to_report <- setdiff(rules_used, xlsx_supported_rules)
+  if (length(rules_to_report) > 0) {
+    warning(paste0(
+      "condformat2excel does not support the following rules: ",
+      paste0(rules_to_report, collapse = ","))
+    )
   }
   xv_cf <- get_xview_and_cf_fields(x)
   xview <- xv_cf[["xview"]]
   cf_fields <- xv_cf[["cf_fields"]]
-
   css_fields <- render_cf_fields_to_css_fields(cf_fields, xview)
-  xlsx::addDataFrame(x = as.data.frame(xview),
-                     sheet = sheet, row.names = FALSE, col.names = TRUE)
-  for (css_key in names(css_fields)) {
-    if (css_key == "background-color") {
-      for (i in seq_len(nrow(xview))) {
-        for (j in seq_len(ncol(xview))) {
-          background_color <- ifelse(css_fields[["background-color"]][i,j] == "",
-                                     NA,
-                                     css_fields[["background-color"]][i,j])
-          if (!is.na(background_color)) {
-            cb <- xlsx::CellBlock.default(sheet, startRow = i + 1, startColumn = j,
-                                          noRows = 1, noColumns = 1, create = FALSE)
-            fill <- xlsx::Fill(backgroundColor = background_color, foregroundColor = background_color)
-            xlsx::CB.setFill(cellBlock = cb,
-                             fill = fill,
-                             rowIndex = 1, colIndex = 1)
-          }
+
+  openxlsx::writeData(wb, sheet_name, as.data.frame(xview),
+                      rowNames = FALSE, colNames = TRUE)
+  created_styles <- list()
+  # For each cell
+  for (i in seq_len(nrow(xview))) {
+    for (j in seq_len(ncol(xview))) {
+      # We build a string that contains all the style information
+      # "hash_background-color:#FF0000;another-css-field:itsvalue;..."
+      fields_vals <- vapply(
+        rlang::set_names(rlang::names2(css_fields)),
+        function(x) css_fields[[x]][i,j],
+        FUN.VALUE = character(1L)
+      )
+      # Remove css keys with either missing values or "":
+      fields_vals <- fields_vals[!is.na(fields_vals) & fields_vals != ""]
+      style_hash <- paste0("hash_",
+        paste(names(fields_vals), fields_vals, sep = ":", collapse = ";")
+      )
+      # If the same style has been defined in another cell, we just use it.
+      # Otherwise we create the style object and we save it:
+      if (style_hash %in% names(created_styles)) {
+        cell_style <- created_styles[[style_hash]]
+      } else {
+        fgFill <- NULL
+        if ("background-color" %in% names(fields_vals)) {
+          fgFill <- fields_vals["background-color"]
         }
+        textDecoration <- NULL
+        if (grepl("bold", fields_vals["font-weight"], fixed = TRUE)) {
+          textDecoration <- c(textDecoration, "bold")
+        }
+        fontColour <- NULL
+        if ("color" %in% names(fields_vals)) {
+          fontColour <- fields_vals["color"]
+        }
+        cell_style <- openxlsx::createStyle(
+          fgFill = fgFill,
+          textDecoration = textDecoration,
+          fontColour = fontColour
+        )
+        created_styles[[style_hash]] <- cell_style
       }
-    } else {
-      warning("The CSS attribute ", css_key, " is not supported in xlsx format")
+      openxlsx::addStyle(wb, sheet_name, style = cell_style, rows = i + 1, cols = j)
     }
   }
+  openxlsx::setColWidths(wb, sheet = sheet_name, cols = seq_len(ncol(xview)), widths = "auto")
   invisible(x)
 }