+++ R/rule_css.R
@@ -7,9 +7,12 @@
 #' @param na.value CSS property value to be used in missing values (e.g. `"grey"`)
 #' @examples
 #' data(iris)
-#' condformat(iris[c(1:5, 51:55, 101:105),]) %>%
+#' cf <- condformat(iris[c(1:5, 51:55, 101:105),]) %>%
 #'   rule_css(Species, expression = ifelse(Species == "setosa", "red", "darkgreen"),
 #'            css_field = "color")
+#' \dontrun{
+#' print(cf)
+#' }
 #' @export
 rule_css <- function(x, columns, expression,
                      css_field,
@@ -33,7 +36,7 @@ rule_css <- function(x, columns, expression,
 }
 
 rule_to_cf_field.rule_css <- function(rule, xfiltered, xview, ...) {
-  columns <- tidyselect::vars_select(colnames(xview), !!! rule[["columns"]])
+  columns <- do.call(tidyselect::vars_select, c(list(colnames(xview)), rule[["columns"]]))
   if (length(columns) == 0) {
     return(NULL)
   }